[chore] set postcss on production
css-loader already use cssnano (on top of postcss) for minification, but we need
to use our custom set of rules for autoprefixer, mqpacker, etc, on production
build. This commit disable the cssnano autoprefixer in favor of custom postcss
ruleset.

It also disable the mergeRules which occurs some bugs with `progress` element
custom style rules.